Split env querying into two files - one for storybook env, and one for component env Grid addon uses storybook env
