+++ .storybook/grid/GridPanel.tsx
@@ -6,7 +6,7 @@ import { API } from '@storybook/api';
 import type { IframePostMessage } from '@guardian/grid-client';
 import { styles } from './gridStyles';
 import { INITIAL_IMAGE_EVENT, IMAGE_SELECTED_EVENT } from './withGrid';
-import { GRID_URL } from '../../src/utils/env';
+import { GRID_URL } from '../utils';
 
 const isValidMessage = (data: IframePostMessage) =>
     data?.crop?.data && data?.image?.data && data?.crop?.data?.assets?.length > 0;
